requirements.txt
fastapi==0.78.0
gunicorn==20.1.0
starlette==0.19.1
uvicorn==0.17.6
python-multipart==0.0.5
opencv-python-headless==4.6.0.66
numpy==1.21.6
lime==0.2.0.1
tensorflow-cpu==2.9.1
Pillow==9.0.1
scikit-image==0.19.2
requests==2.27.1